requirements.txt
anyio==4.12.0
certifi==2025.11.12
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
python-dotenv==1.0.1
python-telegram-bot==20.6
sniffio==1.3.1
